Check if networkSecurityConfig is present and replace it
Because the script only checked if a network security conf file was present with a specific name the script would break for apps where the networkSecurityConfig was already specified in the manifest using a non standard filename. This is because it would add the attribute again.
With this change it checks first if the attribute is present and if so replaces it.